requirements.txt
altgraph==0.17.2
aniso8601==9.0.1
async-generator==1.10
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptocode==0.1
exceptiongroup==1.0.4
Flask==2.2.1
flask-mysql-connector==1.1.0
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
future==0.18.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
pefile==2022.5.30
protobuf==3.20.1
pycparser==2.21
pycryptodomex==3.15.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywin32-ctypes==0.2.0
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
XlsxWriter==3.0.3